Various fixes (hopefully)
RDKit apparently will give hydrogens nan coordinats when adding to a
fragmented methly.  Replace the nans with real numbers.

Code for removing tensors from MolGrid/MolStruct objects.  This wasn't
actually the cause of the memory leak but might still be desirable (can
remove it if necessary).

Untested code for returning the same kind of structure for atom fitting
as the other fitting code.