Output multple samples to a single sdf
Fittings of the real molecule are only output once, to avoid
duplication.  For a stochastic fitting algorithm this might be
undesirable, but when the focus is on generative samples, it is a lot of
unnecessary output (and I would say much of the processing of the
original mol shouldn't even be done, but this only supresses the
output).